requirements.txt
asgiref==3.3.4
Django==3.2.4
django-crispy-forms==1.11.2
django-rest-auth==0.9.5
djangorestframework==3.12.4
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
